Optimized buffer usage

+++ shaders/lib/Settings.glsl
@@ -16,7 +16,7 @@ const int colortex0Format = RG16;
 const int colortex1Format = RGB16;
 const int colortex2Format = RGB16;
 const int colortex3Format = RG16;
-const int colortex4Format = RGBA16;
+const int colortex4Format = RGBA8;
 const int colortex5Format = RGB8;
 */
